change migrations to use bigint fields

+++ rails/db/migrate/20240628174723_create_authors_counts_by_years.rb
@@ -2,10 +2,10 @@ class CreateAuthorsCountsByYears < ActiveRecord::Migration[7.1]
   def change
     create_table :authors_counts_by_years do |t|
       t.references :author, null: false, foreign_key: true
-      t.integer :year, null: false
-      t.integer :works_count
-      t.integer :cited_by_count
-      t.integer :oa_works_count 
+      t.bigint :year, null: false
+      t.bigint :works_count
+      t.bigint :cited_by_count
+      t.bigint :oa_works_count 
 
       t.timestamps
     end
