+++ rails/db/migrate/20240628174736_create_publishers.rb
@@ -4,10 +4,10 @@ class CreatePublishers < ActiveRecord::Migration[7.1]
       t.string :display_name
       t.jsonb :alternate_titles
       t.jsonb :country_codes
-      t.integer :hierarchy_level
+      t.bigint :hierarchy_level
       t.string :parent_publisher
-      t.integer :works_count
-      t.integer :cited_by_count
+      t.bigint :works_count
+      t.bigint :cited_by_count
       t.string :sources_api_url
       
     end
