+++ rails/db/migrate/20240628174737_create_publishers_counts_by_years.rb
@@ -2,10 +2,10 @@ class CreatePublishersCountsByYears < ActiveRecord::Migration[7.1]
   def change
     create_table :publishers_counts_by_years do |t|
       t.references :publisher, null: false, foreign_key: true
-      t.integer :year
-      t.integer :works_count
-      t.integer :cited_by_count
-      t.integer :oa_works_count
+      t.bigint :year
+      t.bigint :works_count
+      t.bigint :cited_by_count
+      t.bigint :oa_works_count
       
     end
   end
