+++ rails/db/migrate/20240628174742_create_works.rb
@@ -6,10 +6,10 @@ class CreateWorks < ActiveRecord::Migration[7.1]
       t.string :doi
       t.string :title
       t.string :display_name
-      t.integer :publication_year
+      t.bigint :publication_year
       t.string :publication_date
       t.string :type
-      t.integer :cited_by_count
+      t.bigint :cited_by_count
       t.boolean :is_retracted
       t.boolean :is_paratext
       t.string :cited_by_api_url
